# backend/docs/swagger.yaml
basePath: /
definitions:
...
  app.GeneralRatingAccessControl:
    properties:
      allowed:
        type: boolean
      role:
        type: string
      user_id:
        type: integer
    type: object
  app.GeneralRatingActivities:
    properties:
      laboratory_works:
        items:
          $ref: '#/definitions/app.GeneralRatingGradedActivity'
        type: array
      lectures:
        items:
          $ref: '#/definitions/app.GeneralRatingAttendanceActivity'
        type: array
      practices:
        items:
          $ref: '#/definitions/app.GeneralRatingGradedActivity'
        type: array
    type: object
  app.GeneralRatingAssessmentSummary:
    properties:
      performance_percent:
        type: number
    type: object
  app.GeneralRatingAttendanceActivity:
    properties:
      attendance_code:
        type: string
      attendance_status:
        type: string
      date:
        type: string
    type: object
  app.GeneralRatingAttendanceSummary:
    properties:
      attendance_percent:
        type: number
    type: object
  app.GeneralRatingConsent:
    properties:
      accepted:
        type: boolean
    type: object
  app.GeneralRatingDepartment:
    properties:
      department_id:
        type: integer
      department_name:
        type: string
    type: object
  app.GeneralRatingGradedActivity:
    properties:
      date:
        type: string
      max_score:
        type: integer
      score:
        type: integer
        x-nullable: true
      title:
        type: string
    type: object
  app.GeneralRatingGroup:
    properties:
      group_id:
        type: integer
      group_name:
        type: string
      students:
        items:
          $ref: '#/definitions/app.GeneralRatingStudent'
        type: array
    type: object
  app.GeneralRatingPagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  app.GeneralRatingPayload:
    properties:
      access_control:
        $ref: '#/definitions/app.GeneralRatingAccessControl'
      departments:
        items:
          $ref: '#/definitions/app.GeneralRatingDepartment'
        type: array
      groups:
        items:
          $ref: '#/definitions/app.GeneralRatingGroup'
        type: array
      pagination:
        $ref: '#/definitions/app.GeneralRatingPagination'
      schema_version:
        type: string
      semester:
        $ref: '#/definitions/app.GeneralRatingSemester'
      subjects:
        items:
          $ref: '#/definitions/app.GeneralRatingSubject'
        type: array
    type: object
  app.GeneralRatingSemester:
    properties:
      semester_id:
        type: integer
      title:
        type: string
    type: object
  app.GeneralRatingStudent:
    properties:
      personal_data_consent:
        $ref: '#/definitions/app.GeneralRatingConsent'
      student_label:
        type: string
      student_ref:
        type: string
      subjects:
        items:
          $ref: '#/definitions/app.GeneralRatingStudentSubject'
        type: array
    type: object
  app.GeneralRatingStudentSubject:
    properties:
      activities:
        $ref: '#/definitions/app.GeneralRatingActivities'
      assessment_summary:
        $ref: '#/definitions/app.GeneralRatingAssessmentSummary'
      attendance_summary:
        $ref: '#/definitions/app.GeneralRatingAttendanceSummary'
      subject_id:
        type: integer
    type: object
  app.GeneralRatingSubject:
    properties:
      department_id:
        type: integer
      department_name:
        type: string
      name:
        type: string
      short_name:
        type: string
      subject_id:
        type: integer
      teacher:
        type: string
      teacher_id:
        type: integer
    type: object
...
  app.Response:
    properties:
      error:
        type: string
      id:
        type: string
      ok:
        type: boolean
      result: {}
    type: object
...
  httpserver.generalRatingResponse:
    properties:
      error:
        example: ""
        type: string
      id:
        example: http-general-rating
        type: string
      ok:
        example: true
        type: boolean
      result:
        $ref: '#/definitions/app.GeneralRatingPayload'
    type: object
...
info:
  contact: {}
  description: Backend API for e-journal, attendance links and invite registration.
  title: EJournal Backend API
  version: "1.0"
paths:
...
  /api/staff/ratings/general:
    get:
      description: Returns semester metadata, role-scoped departments, subjects, groups,
        student consent status, attendance, laboratory/practice grades, and per-subject
        summaries in the standard response envelope.
      parameters:
      - description: Semester ID; defaults to the open semester
        in: query
        name: semester_id
        type: integer
      - description: Group page; defaults to 1
        in: query
        name: page
        type: integer
      - description: Groups per page; defaults to 20, maximum 50
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpserver.generalRatingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: Get source data for the common student rating
      tags:
      - staff
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"